[add] game board class
